build(deps dev): upgrade `@open-draft/test-server` and use its new interface
* Upgrade to the latest available version of `@open-draft/test-server`,
which replaces `async createServer()` closure factory with `new HttpServer()`
and `async HttpServer.prototype.listen()`. That is, previously we awaited
a single function to create the server and start listening, but now we
create the server with a synchronous function and await its listener separately.

* Also (in a rebased version of this commit), upgrade to yet another version
of `@open-draft/test-server` which fixes URL serialization of IPv6 hosts

* Change all affected tests to use this new instantiation method