test(passthrough): warn developer in debug mode that tests will fail
It's useful to run tests in debug mode because it opens a live ("headful")
Playwright browser, e.g. `DEBUG=1 yarn test:integration passthrough` will
run the `passthrough.test.ts` tests in a headful Chromium instance

However, the live debug browser behaves slightly differently from
the headless browser, because there is a navigation request to the
page where the service worker has already been loaded, which causes
the service worker to intercept the initial navigation request.

This is a dangerous trap if you're trying to fix a test by running
it in debug mode. If this is detected, print a friendly warning to
the developer to hopefully save them some time.